requirements.txt
django
djangorestframework-simplejwt
djangorestframework
psycopg2-binary
